LayoutTests/imported/w3c/web-platform-tests/css/css-overscroll-behavior/overscroll-behavior-expected.txt


overscroll-behavior

Tests that overscroll-behavior prevents scroll-propagation in the area and direction as specified.

Make two scrolls on BLUE, in this order: scroll UP (or drag down), then scroll LEFT (or drag right). Scroll (or drag) until nothing is scrolling. Then tap on DONE.
Repeat the same scrolls as in step 1 and then tap on DONE.
Repeat the same scrolls as in step 1 and then tap on DONE.
Make two separate scrolls on GREEN, in this order: scroll UP (or drag down), then scroll LEFT (or drag right). Scroll (or drag) until nothing is scrolling. Then tap on DONE.

FAIL overscroll-behavior prevents scroll-propagation in the area and direction as specified assert_equals: expected 0 but got 100
